+++ src/handler.rs
@@ -1,8 +1,8 @@
 use log::Level::Error as ErrorLevel;
-#[cfg(feature = "ssl")]
-use openssl::ssl::{SslConnectorBuilder, SslMethod, SslStream};
 #[cfg(feature = "nativetls")]
 use native_tls::{TlsConnector, TlsStream as SslStream};
+#[cfg(feature = "ssl")]
+use openssl::ssl::{SslConnector, SslMethod, SslStream};
 use url;
 
 use frame::Frame;
@@ -295,7 +295,7 @@ pub trait Handler {
             Kind::Protocol,
             format!("Unable to parse domain from {}. Needed for SSL.", url),
         ))?;
-        let connector = SslConnectorBuilder::new(SslMethod::tls())
+        let connector = SslConnector::builder(SslMethod::tls())
             .map_err(|e| {
                 Error::new(
                     Kind::Internal,
@@ -317,13 +317,14 @@ pub trait Handler {
             Kind::Protocol,
             format!("Unable to parse domain from {}. Needed for SSL.", url),
         ))?;
-        let connector = TlsConnector::builder().and_then(|builder| builder.build())
-            .map_err(|e| {
-                Error::new(
-                    Kind::Internal,
-                    format!("Failed to upgrade client to SSL: {}", e),
-                )
-            })?;
+
+        let connector = TlsConnector::new().map_err(|e| {
+            Error::new(
+                Kind::Internal,
+                format!("Failed to upgrade client to SSL: {}", e),
+            )
+        })?;
+
         connector.connect(domain, stream).map_err(Error::from)
     }
     /// A method for wrapping a server TcpStream with Ssl Authentication machinery